+++ packages/transports/src/index.ts
@@ -2,3 +2,4 @@ export * from "./otel.js";
 export * from "./http.js";
 export * from "./mongo.js";
 export * from "./loki.js";
+export * from "./prometheus.js";