+++ packages/next/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @aegislog/next
 
+## 0.2.5
+
+### Patch Changes
+
+- Updated dependencies
+  - aegislog@0.2.5
+
 ## 0.2.4
 
 ### Patch Changes